requirements.txt
py-cord
cogwatch
python-dotenv